+++ NoahStolk.SyntaxHighlighter/Piece.cs
@@ -1,21 +1,7 @@
-using System.Web;
-
 namespace NoahStolk.SyntaxHighlighter;
 
-public class Piece
+public sealed class Piece(string code, string type)
 {
-	public Piece(string code, string type)
-	{
-		Code = code;
-		Type = type;
-	}
-
-	public string Code { get; }
-	public string Type { get; }
-
-	// TODO: Move to HtmlWriter.
-	public override string ToString()
-	{
-		return HttpUtility.HtmlEncode(Code);
-	}
+	public string Code { get; } = code;
+	public string Type { get; } = type;
 }