+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 This library uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Changed
+
+- The following types are now sealed:
+  - `Language`
+  - `Piece`
+  - `Style`
+
+### Removed
+
+- Removed custom `Color.ToString()` HTML implementation.
+- Removed custom `Piece.ToString()` HTML implementation.
+
 ## 2.0.0
 
 ### Changed